# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/scan-results": {
            "get": {
                "description": "Get Scan Result List with Offset \u0026 Payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan Results"
                ],
                "summary": "Get Scan Result List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetScanResultListResp"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Scan Result With ScanResult Payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan Results"
                ],
                "summary": "Create Scan Result",
                "parameters": [
                    {
                        "description": "Limit",
                        "name": "limit",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateScanResultResp"
                        }
                    }
                }
            }
        },
        "/v1/scan-results/{id}": {
            "get": {
                "description": "Get Scan Result Detail By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan Results"
                ],
                "summary": "Get Scan Result Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scan Result ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetScanResultDetailResp"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Scan Result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan Results"
                ],
                "summary": "Delete Scan Result",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scan Result ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.DeleteScanResultResp"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Scan Result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan Results"
                ],
                "summary": "Update Scan Result",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scan Result ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateScanResultResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.CreateScanResultResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dtos.DeleteScanResultResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dtos.GetScanResultDetailResp": {
            "type": "object",
            "properties": {
                "scan_result": {
                    "$ref": "#/definitions/dtos.ScanResultResp"
                }
            }
        },
        "dtos.GetScanResultListResp": {
            "type": "object",
            "properties": {
                "scan_results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ScanResultResp"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dtos.ScanResultResp": {
            "type": "object",
            "properties": {
                "findings": {},
                "finished_at": {
                    "type": "string"
                },
                "queued_at": {
                    "type": "string"
                },
                "repository_name": {
                    "type": "string",
                    "example": "test"
                },
                "scanning_at": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dtos.UpdateScanResultResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}